Warning in sphere.cpp fix
git-svn-id: svn://localhost/steel.new/trunk@309 99e027b2-380c-0410-a737-a5784e1a7bde

+++ steel/trunk/sources/objects/sphere.cpp
@@ -113,7 +113,7 @@ void Sphere::ProcessGraph(const GraphEngineInfo &info)
 
 	if(frand()<0.1f && !faces->at(0).triangles->data.empty())
 	{
-		DeleteTriangle(floor((frand()*(faces->at(0).triangles->data.size()-1))));
+		DeleteTriangle((int)floor((frand()*(faces->at(0).triangles->data.size()-1))));
 	}
 	for(unsigned int i = 0; i<normals->data.size(); i++)
 	{
